renamed polaraveraging -> dataops
