+++ docs/doc-config.ini
@@ -1,18 +1,18 @@
 APIDOC
 constants
+dataops
 galaxycluster
 gcdata
 lsst
 modeling
 plotting
-polaraveraging
 utils
 clmm_modeling
 clmm_cosmo
 
 DEMO
 ../examples/demo_generate_mock_cluster.ipynb
-../examples/demo_polaraveraging_functionality.ipynb
+../examples/demo_dataops_functionality.ipynb
 ../examples/demo_modeling_functionality.ipynb
 ../examples/DC2_examples/data_and_model_demo_DC2.ipynb
